requirements.txt
Flask
Flask-PyMongo
python-dotenv
google-generativeai
